worker/requirements.txt
# Python dependencies for steganography worker
Flask==2.3.3
Pillow==10.0.0
numpy==1.24.3
cryptography==41.0.4
python-dotenv==1.0.0
pika==1.3.2
redis==5.0.0
